Move some code to brain-score, add template functions to make this a plug in repo
