+++ candidate_models/__main__.py
@@ -5,9 +5,9 @@ from collections import OrderedDict
 import argparse
 import fire
 
-from candidate_models import score_layers as score_layers_function, score_model as score_model_function, get_activations
 from candidate_models.model_commitments import brain_translated_pool
-from candidate_models.model_commitments.ml_pool import model_layers_pool
+from candidate_models.model_commitments.model_layer_def import model_layers_pool
+from submission import score_layers as score_layers_function, score_model as score_model_function, get_activations
 
 logger = logging.getLogger(__name__)
