+++ setup.py
@@ -14,6 +14,7 @@ requirements = [
     "result_caching @ git+https://github.com/mschrimpf/result_caching",
     "networkx",
     "tqdm",
+    "gitpython",
 
     "cornet @ git+https://github.com/dicarlolab/CORnet",
     "bagnets @ git+https://github.com/mschrimpf/bag-of-local-features-models.git",
