+++ tests/test_behavioral.py
@@ -1,7 +1,8 @@
 import pytest
 from pytest import approx
 
-from candidate_models import brain_translated_pool, score_model
+from candidate_models.model_commitments import brain_translated_pool
+from submission import score_model
 
 
 @pytest.mark.memory_intense
